Add data analysis functionality and License
Added a new script (VW_Datascience.py) that allows to read different files, manage conditions and perform calculations. This code is added to achieve the company's needs in data analysis based on the operations with CSV files and conditions specified. It's going to help to streamline the process and reduce manual work.

Additionally, GPL license (LICENSE.txt) is also added for the repository to promote copying and changes in the software, and protect users' freedom.