Update 'Funktion_B16' and streamline data processing in notebook
Improved the 'Funktion_B16' function in the FINAL notebook to use more explicit conditionals for assign values, which should improve readability and maintainability of the code. Handling of dataframe was simplified by combining several functions and reducing redundant operations. This includes combining various functions for setting the 'C_20' field and adding a 'set_C21_value' function. These changes improve efficiency and clarity of the code.